+++ changelog.md
@@ -7,6 +7,7 @@
        a new Framework ID.
 - CHANGE: !!!! If compose.networks is set, the network mode will be automaticly "user".
        It can be overwritten if compose.networks.driver is set. !!!!
+- ADD: Support for AMD and NVIDIA GPUs.
 
 ## v1.1.3