+++ changelog.md
@@ -9,6 +9,7 @@
        It can be overwritten if compose.networks.driver is set. !!!!
 - ADD: Support for AMD and NVIDIA GPUs.
 - ADD: Mesos Framework GPU Capabilities to get GPU offers from mesos.
+- ADD: Runtime support for docker container (https://github.com/newsnowlabs/runcvm)
 
 ## v1.1.3